Update the test.pl script to handle variable expansions together.
When the variable expansions butted up against one another they would
not be expanded properly. The escape checking broke the regex, and
because the version of perl we run with on RISC OS does not support
the look behind form, we must go to extreme lengths to parse the
expansion.

Adds a capture specification for the command output.