Add some tests and invocation code for running on RISC OS.
The test code here is able to run on RISC OS to check that the behaviour
is correct. This meant I needed to write a small tool to capture the
error reports through OS_GenerateError. The capture tool isn't great, but
it gives me the opportunity to use it in a way very similar to that of
the Pyromaniac environment.

I had to update the test.pl tool too, but that's not too terrible - it's
intended to be used in both RISC OS and POSIX environments. Largely it's
compatibility for the ancient perl I'm using on RISC OS.