Add some actual tests of the OS_ReadArgs call, and VDU system.
OS_ReadArgs is a relatively complex call, and it's useful to be able to
show that it's working right. There's some discrepencies on how the
calls return values with what RISC OS Pyromaniac does, which is kinda
interesting.

The VDU system has a simple test that the mode variables are returned,
and a few sources updated with fixes and changes from Pyromaniac.