Update test code from Pyromaniac.
The test code files that have changed in Pyromaniac have been updated
here so that we have consistency between the two.

Changing environment shouldn't change the registers that are not used
by the environment handler. This is now checked.

OS_Write0 returns the width in bytes, without the terminator. This had
been wrong on Pyromaniac.

Kay handling not tests the caps lock behaviour.

Variable reading now uses the context variables to enumerate all the
variables that match.