Add tests for Debugger, as used in Pyromaniac.
The tests for Debugger just disassemble a bunch of instructions.
That's pretty much it, and then we check that they match up to what
we are expecting. The output will almost certainly differ from that
produced by the Classic debugger, but the important thing is that
the output is consistent (and obviously correct).