Update keyv tests to add *FX 4 configuration; add s._args_common.
The Cursor and F-key configuration is now able to be exercised with
the keyv tests; this also allows us to change *FX4 values to exercise
these.

The s._args_common, which is used in a lot of the modern Pyromaniac
tests has also been imported to make it possible to use this test
and probably others in the future.